requirements.txt
gradio==5.3.0
openai>=1.79.0
fal-client>=0.6.0
Pillow>=10.0.0
typing-extensions>=4.7.0 